[fa] DataBase powered Game Data 数据库驱动的游戏数据系统 谁家宝贝的早八上人体解剖啊南医大

+++ src/main/java/com/vbuser/database/operate/New.java
@@ -22,6 +22,10 @@ public class New {
     public static void createTable(String name, String[] columns, File dataBase) {
         // 创建表文件并写入表头
         File tableFile = new File(new File(dataBase, "tables"), name + ".txt");
+        if(tableFile.exists()){
+            System.out.println("[] Skipped as the table has existed.");
+            return;
+        }
         try (FileWriter writer = new FileWriter(tableFile, true)) {
             for (int i = 0; i < columns.length; i++) {
                 writer.write(columns[i]);
